+++ CHANGELOG.md
@@ -7,6 +7,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.6.4] - 2023-09-24
+
+### Fixed
+
+- unary minus operator
+- engine update
+
+### Added
+
+- bind decorator
+- unary plus operator
+- conversion to number
+
 ## [2.6.3] - 2023-09-22
 
 ### Fixed
@@ -200,7 +213,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Reserved words are no longer in Generator's constructor.
 
-[unreleased]: https://github.com/tomas-wrobel/scrap/compare/v2.6.3...HEAD
+[unreleased]: https://github.com/tomas-wrobel/scrap/compare/v2.6.4...HEAD
+[2.6.4]: https://github.com/tomas-wrobel/scrap/releases/tag/v2.6.4
 [2.6.3]: https://github.com/tomas-wrobel/scrap/releases/tag/v2.6.3
 [2.6.2]: https://github.com/tomas-wrobel/scrap/releases/tag/v2.6.2
 [2.6.1]: https://github.com/tomas-wrobel/scrap/releases/tag/v2.6.1